Speedrunning duplicate code
