+++ src/main/scala/dpla/ingestion3/harvesters/file/DlgFileHarvester.scala
@@ -33,19 +33,6 @@ class DlgFileHarvester(
 
   protected val extractor = new DlgFileExtractor()
 
-  /** Loads .zip files
-    *
-    * @param file
-    *   File to parse
-    * @return
-    *   ZipInputstream of the zip contents
-    */
-  def getInputStream(file: File): Option[ZipInputStream] =
-    file.getName match {
-      case zipName if zipName.endsWith("zip") =>
-        Some(new ZipInputStream(new FileInputStream(file)))
-      case _ => None
-    }
 
   /** Parses JValue to extract item local item id and renders compact full
     * record
@@ -118,7 +105,7 @@ class DlgFileHarvester(
     inFiles
       .listFiles(zipFilter)
       .foreach(inFile => {
-        val inputStream: ZipInputStream = getInputStream(inFile)
+        val inputStream: ZipInputStream = FileHarvester.getZipInputStream(inFile)
           .getOrElse(
             throw new IllegalArgumentException("Couldn't load ZIP files.")
           )
