+++ src/main/scala/dpla/ingestion3/harvesters/file/VtFileHarvester.scala
@@ -3,9 +3,9 @@ package dpla.ingestion3.harvesters.file
 import java.io.{File, FileInputStream}
 import java.util.zip.ZipInputStream
 import dpla.ingestion3.confs.i3Conf
+import dpla.ingestion3.harvesters.Harvester
 import dpla.ingestion3.mappers.utils.XmlExtractor
 import org.apache.commons.io.IOUtils
-
 import org.apache.spark.sql.{DataFrame, SparkSession}
 import dpla.ingestion3.harvesters.file.FileFilters.zipFilter
 import dpla.ingestion3.model.AVRO_MIME_XML
@@ -27,21 +27,6 @@ class VtFileHarvester(
 
   protected val extractor = new VtFileExtractor
 
-  /** Loads .zip files
-    *
-    * @param file
-    *   File to parse
-    * @return
-    *   ZipInputstream of the zip contents
-    *
-    */
-  def getInputStream(file: File): Option[ZipInputStream] =
-    file.getName match {
-      case zipName if zipName.endsWith("zip") =>
-        Some(new ZipInputStream(new FileInputStream(file)))
-      case _ => None
-    }
-
   /** Main logic for handling individual entries in the zip.
     *
     * @param zipResult
@@ -61,7 +46,7 @@ class VtFileHarvester(
           } match {
             case Success(xml) =>
               val id: String = zipResult.entryName
-              val outputXml: String = xmlToString(xml)
+              val outputXml: String = Harvester.xmlToString(xml)
               val item: ParsedResult = ParsedResult(id, outputXml)
               writeOut(unixEpoch, item)
               1
@@ -80,7 +65,7 @@ class VtFileHarvester(
     inFiles
       .listFiles(zipFilter)
       .foreach(inFile => {
-        val inputStream = getInputStream(inFile)
+        val inputStream = FileHarvester.getZipInputStream(inFile)
           .getOrElse(
             throw new IllegalArgumentException("Couldn't load ZIP files.")
           )
@@ -100,15 +85,5 @@ class VtFileHarvester(
     // Read harvested data into Spark DataFrame and return.
     spark.read.format("avro").load(tmpOutStr)
   }
-
-  /** Converts a Node to an xml string
-    *
-    * @param node
-    *   The root of the tree to write to a string
-    * @return
-    *   a String containing xml
-    */
-  def xmlToString(node: Node): String =
-    Utility.serialize(node, minimizeTags = MinimizeMode.Always).toString
 
 }
